+++ integration/system/disk_usage_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/moby/moby/api/types/build"
 	containertypes "github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/image"
-	"github.com/moby/moby/api/types/system"
 	"github.com/moby/moby/api/types/volume"
 	"github.com/moby/moby/client"
 	"github.com/moby/moby/v2/integration/internal/container"
@@ -30,78 +29,110 @@ func TestDiskUsage(t *testing.T) {
 	defer d.Stop(t)
 	apiClient := d.NewClientT(t)
 
-	var stepDU system.DiskUsage
+	var stepDU client.DiskUsageResult
 	for _, step := range []struct {
 		doc  string
-		next func(t *testing.T, prev system.DiskUsage) system.DiskUsage
+		next func(t *testing.T, prev client.DiskUsageResult) client.DiskUsageResult
 	}{
 		{
 			doc: "empty",
-			next: func(t *testing.T, _ system.DiskUsage) system.DiskUsage {
-				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{})
+			next: func(t *testing.T, _ client.DiskUsageResult) client.DiskUsageResult {
+				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{
+					Images:     true,
+					Containers: true,
+					BuildCache: true,
+					Volumes:    true,
+					Verbose:    true,
+				})
 				assert.NilError(t, err)
 
 				expectedLayersSize := int64(0)
 				// TODO: Investigate https://github.com/moby/moby/issues/47119
 				// Make 4096 (block size) also a valid value for zero usage.
 				if testEnv.UsingSnapshotter() && testEnv.IsRootless() {
-					if du.LayersSize == 4096 {
-						expectedLayersSize = du.LayersSize
+					if du.Images.TotalSize == 4096 {
+						expectedLayersSize = 4096
 					}
 				}
 
-				assert.DeepEqual(t, du, system.DiskUsage{
-					LayersSize: expectedLayersSize,
-					Images:     []*image.Summary{},
-					Containers: []*containertypes.Summary{},
-					Volumes:    []*volume.Volume{},
-					BuildCache: []*build.CacheRecord{},
+				assert.DeepEqual(t, du, client.DiskUsageResult{
+					Containers: client.ContainersDiskUsage{
+						Items: []containertypes.Summary{},
+					},
+					Images: client.ImagesDiskUsage{
+						TotalSize: expectedLayersSize,
+						Items:     []image.Summary{},
+					},
+					BuildCache: client.BuildCacheDiskUsage{
+						Items: []build.CacheRecord{},
+					},
+					Volumes: client.VolumesDiskUsage{
+						Items: []volume.Volume{},
+					},
 				})
 				return du
 			},
 		},
 		{
 			doc: "after LoadBusybox",
-			next: func(t *testing.T, _ system.DiskUsage) system.DiskUsage {
+			next: func(t *testing.T, _ client.DiskUsageResult) client.DiskUsageResult {
 				d.LoadBusybox(ctx, t)
 
-				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{})
+				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{
+					Images:     true,
+					Containers: true,
+					BuildCache: true,
+					Volumes:    true,
+					Verbose:    true,
+				})
 				assert.NilError(t, err)
-				assert.Assert(t, du.LayersSize > 0)
-				assert.Equal(t, len(du.Images), 1)
-				assert.Equal(t, len(du.Images[0].RepoTags), 1)
-				assert.Check(t, is.Equal(du.Images[0].RepoTags[0], "busybox:latest"))
+
+				assert.Assert(t, du.Images.TotalSize > 0)
+				assert.Equal(t, len(du.Images.Items), 1)
+				assert.Equal(t, len(du.Images.Items[0].RepoTags), 1)
+				assert.Check(t, is.Equal(du.Images.Items[0].RepoTags[0], "busybox:latest"))
 
 				// Image size is layer size + content size. Content size is included in layers size.
-				assert.Equal(t, du.Images[0].Size, du.LayersSize)
+				assert.Equal(t, du.Images.Items[0].Size, du.Images.TotalSize)
 
 				return du
 			},
 		},
 		{
 			doc: "after container.Run",
-			next: func(t *testing.T, prev system.DiskUsage) system.DiskUsage {
+			next: func(t *testing.T, prev client.DiskUsageResult) client.DiskUsageResult {
 				cID := container.Run(ctx, t, apiClient)
 
-				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{})
+				du, err := apiClient.DiskUsage(ctx, client.DiskUsageOptions{
+					Images:     true,
+					Containers: true,
+					BuildCache: true,
+					Volumes:    true,
+					Verbose:    true,
+				})
 				assert.NilError(t, err)
-				assert.Equal(t, len(du.Containers), 1)
-				assert.Equal(t, len(du.Containers[0].Names), 1)
-				assert.Assert(t, len(prev.Images) > 0)
-				assert.Check(t, du.Containers[0].Created >= prev.Images[0].Created)
+
+				assert.Equal(t, du.Containers.ActiveContainers, int64(1))
+				assert.Equal(t, du.Containers.TotalContainers, int64(1))
+				assert.Equal(t, len(du.Containers.Items), 1)
+				assert.Equal(t, len(du.Containers.Items[0].Names), 1)
+				assert.Assert(t, len(prev.Images.Items) > 0)
+				assert.Check(t, du.Containers.Items[0].Created >= prev.Images.Items[0].Created)
 
 				// Additional container layer could add to the size
-				assert.Check(t, du.LayersSize >= prev.LayersSize)
+				assert.Check(t, du.Images.TotalSize >= prev.Images.TotalSize)
 
-				assert.Equal(t, len(du.Images), 1)
-				assert.Equal(t, du.Images[0].Containers, prev.Images[0].Containers+1)
+				assert.Equal(t, du.Images.ActiveImages, int64(1))
+				assert.Equal(t, du.Images.TotalImages, int64(1))
+				assert.Equal(t, len(du.Images.Items), 1)
+				assert.Equal(t, du.Images.Items[0].Containers, prev.Images.Items[0].Containers+1)
 
-				assert.Check(t, is.Equal(du.Containers[0].ID, cID))
-				assert.Check(t, is.Equal(du.Containers[0].Image, "busybox"))
-				assert.Check(t, is.Equal(du.Containers[0].ImageID, prev.Images[0].ID))
+				assert.Check(t, is.Equal(du.Containers.Items[0].ID, cID))
+				assert.Check(t, is.Equal(du.Containers.Items[0].Image, "busybox"))
+				assert.Check(t, is.Equal(du.Containers.Items[0].ImageID, prev.Images.Items[0].ID))
 
 				// ImageManifestDescriptor should NOT be populated.
-				assert.Check(t, is.Nil(du.Containers[0].ImageManifestDescriptor))
+				assert.Check(t, is.Nil(du.Containers.Items[0].ImageManifestDescriptor))
 
 				return du
 			},
@@ -114,143 +145,153 @@ func TestDiskUsage(t *testing.T) {
 			for _, tc := range []struct {
 				doc      string
 				options  client.DiskUsageOptions
-				expected system.DiskUsage
+				expected client.DiskUsageResult
 			}{
 				{
 					doc: "container types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ContainerObject,
-						},
+						Containers: true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
+					expected: client.DiskUsageResult{
 						Containers: stepDU.Containers,
+						Images:     client.ImagesDiskUsage{Items: []image.Summary{}},
+						BuildCache: client.BuildCacheDiskUsage{Items: []build.CacheRecord{}},
+						Volumes:    client.VolumesDiskUsage{Items: []volume.Volume{}},
 					},
 				},
 				{
 					doc: "image types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ImageObject,
-						},
+						Images:  true,
+						Verbose: true,
 					},
-					expected: system.DiskUsage{
-						LayersSize: stepDU.LayersSize,
+					expected: client.DiskUsageResult{
+						Containers: client.ContainersDiskUsage{Items: []containertypes.Summary{}},
 						Images:     stepDU.Images,
+						BuildCache: client.BuildCacheDiskUsage{Items: []build.CacheRecord{}},
+						Volumes:    client.VolumesDiskUsage{Items: []volume.Volume{}},
 					},
 				},
 				{
 					doc: "volume types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.VolumeObject,
-						},
+						Volumes: true,
+						Verbose: true,
 					},
-					expected: system.DiskUsage{
-						Volumes: stepDU.Volumes,
+					expected: client.DiskUsageResult{
+						Containers: client.ContainersDiskUsage{Items: []containertypes.Summary{}},
+						Images:     client.ImagesDiskUsage{Items: []image.Summary{}},
+						BuildCache: client.BuildCacheDiskUsage{Items: []build.CacheRecord{}},
+						Volumes:    stepDU.Volumes,
 					},
 				},
 				{
 					doc: "build-cache types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.BuildCacheObject,
-						},
+						BuildCache: true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
+					expected: client.DiskUsageResult{
+						Containers: client.ContainersDiskUsage{Items: []containertypes.Summary{}},
+						Images:     client.ImagesDiskUsage{Items: []image.Summary{}},
 						BuildCache: stepDU.BuildCache,
+						Volumes:    client.VolumesDiskUsage{Items: []volume.Volume{}},
 					},
 				},
 				{
 					doc: "container, volume types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ContainerObject,
-							system.VolumeObject,
-						},
+						Containers: true,
+						Volumes:    true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
+					expected: client.DiskUsageResult{
 						Containers: stepDU.Containers,
+						Images:     client.ImagesDiskUsage{Items: []image.Summary{}},
+						BuildCache: client.BuildCacheDiskUsage{Items: []build.CacheRecord{}},
 						Volumes:    stepDU.Volumes,
 					},
 				},
 				{
 					doc: "image, build-cache types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ImageObject,
-							system.BuildCacheObject,
-						},
+						Images:     true,
+						BuildCache: true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
-						LayersSize: stepDU.LayersSize,
+					expected: client.DiskUsageResult{
+						Containers: client.ContainersDiskUsage{Items: []containertypes.Summary{}},
 						Images:     stepDU.Images,
 						BuildCache: stepDU.BuildCache,
+						Volumes:    client.VolumesDiskUsage{Items: []volume.Volume{}},
 					},
 				},
 				{
 					doc: "container, volume, build-cache types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ContainerObject,
-							system.VolumeObject,
-							system.BuildCacheObject,
-						},
+						Containers: true,
+						BuildCache: true,
+						Volumes:    true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
+					expected: client.DiskUsageResult{
 						Containers: stepDU.Containers,
-						Volumes:    stepDU.Volumes,
+						Images: client.ImagesDiskUsage{
+							Items: []image.Summary{},
+						},
 						BuildCache: stepDU.BuildCache,
+						Volumes:    stepDU.Volumes,
 					},
 				},
 				{
 					doc: "image, volume, build-cache types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ImageObject,
-							system.VolumeObject,
-							system.BuildCacheObject,
-						},
+						Images:     true,
+						BuildCache: true,
+						Volumes:    true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
-						LayersSize: stepDU.LayersSize,
+					expected: client.DiskUsageResult{
+						Containers: client.ContainersDiskUsage{
+							Items: []containertypes.Summary{},
+						},
 						Images:     stepDU.Images,
-						Volumes:    stepDU.Volumes,
 						BuildCache: stepDU.BuildCache,
+						Volumes:    stepDU.Volumes,
 					},
 				},
 				{
 					doc: "container, image, volume types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ContainerObject,
-							system.ImageObject,
-							system.VolumeObject,
-						},
+						Containers: true,
+						Images:     true,
+						Volumes:    true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
-						LayersSize: stepDU.LayersSize,
+					expected: client.DiskUsageResult{
 						Containers: stepDU.Containers,
 						Images:     stepDU.Images,
-						Volumes:    stepDU.Volumes,
+						BuildCache: client.BuildCacheDiskUsage{
+							Items: []build.CacheRecord{},
+						},
+						Volumes: stepDU.Volumes,
 					},
 				},
 				{
 					doc: "container, image, volume, build-cache types",
 					options: client.DiskUsageOptions{
-						Types: []system.DiskUsageObject{
-							system.ContainerObject,
-							system.ImageObject,
-							system.VolumeObject,
-							system.BuildCacheObject,
-						},
+						Containers: true,
+						Images:     true,
+						BuildCache: true,
+						Volumes:    true,
+						Verbose:    true,
 					},
-					expected: system.DiskUsage{
-						LayersSize: stepDU.LayersSize,
+					expected: client.DiskUsageResult{
 						Containers: stepDU.Containers,
 						Images:     stepDU.Images,
-						Volumes:    stepDU.Volumes,
 						BuildCache: stepDU.BuildCache,
+						Volumes:    stepDU.Volumes,
 					},
 				},
 			} {